+++ internal/orderedmap/orderedmap.go
@@ -1,0 +1,263 @@
+/*
+ This module is a modified version of the original work.
+ Original work can be found at:
+          https://github.com/wk8/go-ordered-map
+
+ The license provided (copy_of_license.md) covers the files
+ within this directory. In addition, the changes or modifications
+ done are described in changes.md.
+ I do not claim ownership or creation of this module. All
+ credit should be given to the original author.
+*/
+
+// Package orderedmap implements an ordered map, i.e. a map that also keeps track of
+// the order in which keys were inserted.
+
+package orderedmap
+
+import (
+	"fmt"
+
+	list "github.com/bahlo/generic-list-go"
+)
+
+type Pair[K comparable, V any] struct {
+	Key   K
+	Value V
+
+	element *list.Element[*Pair[K, V]]
+}
+
+type OrderedMap[K comparable, V any] struct {
+	pairs map[K]*Pair[K, V]
+	list  *list.List[*Pair[K, V]]
+}
+
+// New creates a new OrderedMap.
+func New[K comparable, V any]() *OrderedMap[K, V] {
+	return &OrderedMap[K, V]{
+		pairs: make(map[K]*Pair[K, V]),
+		list:  list.New[*Pair[K, V]](),
+	}
+}
+
+// Clone creates a new OrderedMap that has the same mappings as the current map.
+func (om *OrderedMap[K, V]) Clone() *OrderedMap[K, V] {
+	ret := &OrderedMap[K, V]{
+		pairs: make(map[K]*Pair[K, V]),
+		list:  list.New[*Pair[K, V]](),
+	}
+	for pair := om.Oldest(); pair != nil; pair = pair.Next() {
+		ret.setNewPair(pair, true)
+	}
+	return ret
+}
+
+// Copy is an alias for Clone.
+func (om *OrderedMap[K, V]) Copy() *OrderedMap[K, V] {
+	return om.Clone()
+}
+
+// Reverse reverses the order of the mappings within a map.
+func (om *OrderedMap[K, V]) Reverse() {
+	curr := om.list.Front()
+	prev, next := curr, curr.Next()
+	for curr != nil {
+		next = curr.Next()
+		om.list.MoveBefore(curr, prev)
+		prev = curr
+		curr = next
+	}
+}
+
+// Get looks for the given key, and returns the value associated with it,
+// or V's nil value if not found. The boolean it returns says whether the key is present in the map.
+func (om *OrderedMap[K, V]) Get(key K) (val V, present bool) {
+	if pair, present := om.pairs[key]; present {
+		return pair.Value, true
+	}
+	return
+}
+
+// Load is an alias for Get, mostly to present an API similar to `sync.Map`'s.
+func (om *OrderedMap[K, V]) Load(key K) (V, bool) {
+	return om.Get(key)
+}
+
+// GetPair looks for the given key, and returns the pair associated with it,
+// or nil if not found. The Pair struct can then be used to iterate over the ordered map
+// from that point, either forward or backward.
+func (om *OrderedMap[K, V]) GetPair(key K) *Pair[K, V] {
+	return om.pairs[key]
+}
+
+// Set sets the key-value pair, and returns what `Get` would have returned
+// on that key prior to the call to `Set`.
+func (om *OrderedMap[K, V]) Set(key K, value V, old bool) (val V, present bool) {
+	if pair, present := om.pairs[key]; present {
+		oldValue := pair.Value
+		pair.Value = value
+		if !old {
+			om.list.MoveToFront(pair.element)
+		}
+		return oldValue, true
+	}
+	pair := &Pair[K, V]{
+		Key:   key,
+		Value: value,
+	}
+	om.setNewPair(pair, old)
+	return
+}
+
+// Store is an alias for Set, mostly to present an API similar to `sync.Map`'s.
+func (om *OrderedMap[K, V]) Store(key K, value V, old bool) (V, bool) {
+	return om.Set(key, value, old)
+}
+
+// Delete removes the key-value pair, and returns what `Get` would have returned
+// on that key prior to the call to `Delete`.
+func (om *OrderedMap[K, V]) Delete(key K) (val V, present bool) {
+	if pair, present := om.pairs[key]; present {
+		om.list.Remove(pair.element)
+		delete(om.pairs, key)
+		return pair.Value, true
+	}
+	return
+}
+
+func (om *OrderedMap[K, V]) setNewPair(pair *Pair[K, V], old bool) {
+	if old {
+		pair.element = om.list.PushBack(pair)
+	} else {
+		pair.element = om.list.PushFront(pair)
+	}
+	om.pairs[pair.Key] = pair
+}
+
+// Len returns the length of the ordered map.
+func (om *OrderedMap[K, V]) Len() int {
+	return len(om.pairs)
+}
+
+// Size is an alias for Len.
+func (om *OrderedMap[K, V]) Size() int {
+	return om.Len()
+}
+
+// Oldest returns a pointer to the oldest pair. It's meant to be used to iterate on the ordered map's
+// pairs from the oldest to the newest, e.g.:
+// for pair := orderedMap.Oldest(); pair != nil; pair = pair.Next() { fmt.Printf("%v => %v\n", pair.Key, pair.Value) }
+func (om *OrderedMap[K, V]) Oldest() *Pair[K, V] {
+	return listElementToPair(om.list.Front())
+}
+
+// First is an alias for Oldest, as sometimes it can be more clear.
+func (om *OrderedMap[K, V]) First() *Pair[K, V] {
+	return om.Oldest()
+}
+
+// Front is an alias for Oldest, as sometimes it can be more clear.
+func (om *OrderedMap[K, V]) Front() *Pair[K, V] {
+	return om.Oldest()
+}
+
+// Newest returns a pointer to the newest pair. It's meant to be used to iterate on the ordered map's
+// pairs from the newest to the oldest, e.g.:
+// for pair := orderedMap.Oldest(); pair != nil; pair = pair.Next() { fmt.Printf("%v => %v\n", pair.Key, pair.Value) }
+func (om *OrderedMap[K, V]) Newest() *Pair[K, V] {
+	return listElementToPair(om.list.Back())
+}
+
+// Last is an alias for Newest, as sometimes it can be more clear.
+func (om *OrderedMap[K, V]) Last() *Pair[K, V] {
+	return om.Newest()
+}
+
+// Back is an alias for Newest, as sometimes it can be more clear.
+func (om *OrderedMap[K, V]) Back() *Pair[K, V] {
+	return om.Newest()
+}
+
+// Next returns a pointer to the next pair.
+func (p *Pair[K, V]) Next() *Pair[K, V] {
+	return listElementToPair(p.element.Next())
+}
+
+// Prev returns a pointer to the previous pair.
+func (p *Pair[K, V]) Prev() *Pair[K, V] {
+	return listElementToPair(p.element.Prev())
+}
+
+func listElementToPair[K comparable, V any](element *list.Element[*Pair[K, V]]) *Pair[K, V] {
+	if element == nil {
+		return nil
+	}
+	return element.Value
+}
+
+// KeyNotFoundError may be returned by functions in this package when they're called with keys that are not present
+// in the map.
+type KeyNotFoundError[K comparable] struct {
+	MissingKey K
+}
+
+func (e *KeyNotFoundError[K]) Error() string {
+	return fmt.Sprintf("missing key: %v", e.MissingKey)
+}
+
+// MoveAfter moves the value associated with key to its new position after the one associated with markKey.
+// Returns an error iff key or markKey are not present in the map.
+func (om *OrderedMap[K, V]) MoveAfter(key, markKey K) error {
+	elements, err := om.getElements(key, markKey)
+	if err != nil {
+		return err
+	}
+	om.list.MoveAfter(elements[0], elements[1])
+	return nil
+}
+
+// MoveBefore moves the value associated with key to its new position before the one associated with markKey.
+// Returns an error iff key or markKey are not present in the map.
+func (om *OrderedMap[K, V]) MoveBefore(key, markKey K) error {
+	elements, err := om.getElements(key, markKey)
+	if err != nil {
+		return err
+	}
+	om.list.MoveBefore(elements[0], elements[1])
+	return nil
+}
+
+func (om *OrderedMap[K, V]) getElements(keys ...K) ([]*list.Element[*Pair[K, V]], error) {
+	elements := make([]*list.Element[*Pair[K, V]], len(keys))
+	for i, k := range keys {
+		pair, present := om.pairs[k]
+		if !present {
+			return nil, &KeyNotFoundError[K]{k}
+		}
+		elements[i] = pair.element
+	}
+	return elements, nil
+}
+
+// MoveToBack moves the value associated with key to the back of the ordered map.
+// Returns an error iff key is not present in the map.
+func (om *OrderedMap[K, V]) MoveToBack(key K) error {
+	pair, present := om.pairs[key]
+	if !present {
+		return &KeyNotFoundError[K]{key}
+	}
+	om.list.MoveToBack(pair.element)
+	return nil
+}
+
+// MoveToFront moves the value associated with key to the front of the ordered map.
+// Returns an error iff key is not present in the map.
+func (om *OrderedMap[K, V]) MoveToFront(key K) error {
+	pair, present := om.pairs[key]
+	if !present {
+		return &KeyNotFoundError[K]{key}
+	}
+	om.list.MoveToFront(pair.element)
+	return nil
+}